+++ spec/twirp/protoc_plugin/descriptor_ext/file_descriptor_proto_ext_spec.rb
@@ -24,4 +24,27 @@ RSpec.describe Google::Protobuf::FileDescriptorProto do
       expect(proto_file.relative_ruby_protobuf_name).to eq("hello_pb")
     end
   end
+
+  describe "#has_service?" do
+    it "returns false when there are no services" do
+      proto_file = Google::Protobuf::FileDescriptorProto.new(name: "dummy.proto")
+      expect(proto_file.has_service?).to eq(false)
+    end
+
+    it "returns true when there there is one service" do
+      proto_file = Google::Protobuf::FileDescriptorProto.new(
+        name: "dummy.proto",
+        service: [Google::Protobuf::ServiceDescriptorProto.new]
+      )
+      expect(proto_file.has_service?).to eq(true)
+    end
+
+    it "returns true when there there are multiple services" do
+      proto_file = Google::Protobuf::FileDescriptorProto.new(
+        name: "dummy.proto",
+        service: [Google::Protobuf::ServiceDescriptorProto.new, Google::Protobuf::ServiceDescriptorProto.new]
+      )
+      expect(proto_file.has_service?).to eq(true)
+    end
+  end
 end